Renamed component and removed withContent
